Remove files not directly needed by method or examples, rename utils folder in dg.
