+++ sfepy/discrete/dg/utils/inits_consts.py
